Preforms checks before sending commands to an XBMC server
Checks that there is an active XBMC Config active. If there is will
check that the correct API is loaded. If it is it will load execute the
command, otherwise it will try to load the API. If the API loads it
will send the command.

The current Timeout is ~10secs.